requirements.txt
# Web framework
Flask==3.0.0
Werkzeug==3.0.1

# SMB3 client и зависимости
smbprotocol==1.12.0
pyspnego==0.10.2
cryptography>=3.4.8
ntlm-auth>=1.5.0

# HTTP requests  
requests==2.31.0

# Environment variables
python-dotenv==1.0.0

# Additional dependencies для SMB3
six>=1.16.0
urllib3==2.1.0
